+++ src/vatensor/rearrange.hpp
@@ -53,6 +53,7 @@ namespace va {
 
 	std::shared_ptr<VArray> squeeze(const std::shared_ptr<VArray>& vdata);
 	std::shared_ptr<VArray> squeeze(const std::shared_ptr<VArray>& vdata, const axes_type& axes);
+	std::shared_ptr<VArray> broadcast_to(const VArray& varray, const shape_type& target_shape);
 	std::shared_ptr<VArray> expand_dims(const VArray& varray, std::ptrdiff_t axis);
 }